Restructure the flow, extracting flow steps into separate functions for legibility. The main flow now fits on one screen, and each function more or less does the same. This also allowed some re-arranging and compression.
